Move zip methods to new module
